requirements.txt
streamlit
pandas
joblib
pillow
scikit-learn
